fix: prevent gateway spawning during precompilation, fix precision/leverage type errors
Root cause: importing a module should not spawn external processes.
Added Base.generating_output() guards in three places to skip gateway
operations during precompilation:
  - _ensure_gateway_running() in CcxtGateway/rest.jl
  - Exchange() constructor spawn logic in ExchangeTypes/exchange.jl
  - sandbox!() in Exchanges/constructors.jl

Additionally fixed cascading type errors that surfaced once gateway
spawn was skipped during precompilation:
  - tickers.jl: market_precision now always returns Float64 values
  - tickers.jl: _minmax_pair wraps something() chain with Float64()
  - module.jl: positions() handles empty leverage_tiers gracefully
  - constructors.jl: AssetInstance passes NamedTuple prec, not raw
    precision tuple, to inner constructor
  - positions.jl: tier! guards against nothing assignment when no
    matching leverage tier is found

+++ Exchanges/src/tickers.jl
@@ -312,10 +312,10 @@ $(TYPEDSIGNATURES)
 """
 function market_precision(pair, exc)
     mkt = get(exc.markets, pair, nothing)
-    mkt === nothing && return (default_amount_precision(exc), default_price_precision(exc))
+    mkt === nothing && return (Float64(default_amount_precision(exc)), Float64(default_price_precision(exc)))
     prec = get(mkt, "precision", Dict())
-    amt = decimal_to_size(_get_precision(exc, mkt, "amount"), exc.precision; exc)
-    prc = decimal_to_size(_get_precision(exc, mkt, "price"), exc.precision; exc)
+    amt = Float64(decimal_to_size(_get_precision(exc, mkt, "amount"), exc.precision; exc))
+    prc = Float64(decimal_to_size(_get_precision(exc, mkt, "price"), exc.precision; exc))
     (amt, prc)
 end
